requirements.txt
# Base Math & Data (Force these first)
numpy>=2.1.0
pandas>=2.2.0

# Voice & Telephony
livekit>=1.1.3
livekit-agents>=1.5.1
twilio>=9.0.0

# LLM & Orchestration (Updated for NumPy 2.x compatibility)
langchain>=0.3.10
langchain-community>=0.3.10
langgraph>=0.2.50
langchain-groq>=0.2.1
groq>=0.11.0

# Vector Store & RAG
qdrant-client>=1.12.0
sentence-transformers>=3.3.0

# Web & Utils
streamlit>=1.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Free Voice Processing
openai-whisper>=20231117
gtts>=2.5.0
pyttsx3  # Optional: offline TTS backup

gunicorn>=21.0.0